requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
easyocr==1.4.2
elastic-transport==8.1.2
elasticsearch==8.1.3
elasticsearch-async==6.2.0
Flask==2.1.2
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
imageio==2.18.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
networkx==2.8
numpy==1.22.3
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-bidi==0.4.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
tifffile==2022.4.28
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
